# 10 times faster compiling speed