--- a/package.json
+++ b/package.json
@@ -1,6 +1,6 @@
 {
   "name": "flowbite-svelte",
-  "version": "0.32.1",
+  "version": "0.32.2",
   "description": "Flowbite components for Svelte",
   "author": {
     "name": "Shinichi Okada",