This update allows users who use a small number of components to shorten the compiling time:

```
import Alert from 'flowbite-svelte/Alert.svelte'
```

|package |        files |    time|     avg|
|--|--|--|--|
|flowbite-svelte|     4|   23.8ms |  6.0ms|

If you use a normal import:

```
import { Alert } from 'flowbite-svelte';
```

|package |        files |    time|     avg|
|--|--|--|--|
|flowbite-svelte|   145 |  0.23s|   1.6ms|

Ten times faster.

The main change is to add individual exports to package.json.

```
 "exports": {
    "./Accordion.svelte": {
      "types": "./dist/accordions/Accordion.svelte.d.ts",
      "svelte": "./dist/accordions/Accordion.svelte"
    },
    "./AccordionItem.svelte": {
      "types": "./dist/accordions/AccordionItem.svelte.d.ts",
      "svelte": "./dist/accordions/AccordionItem.svelte"
    },
...
```

A new script, add-exports-to-package-json.js, is added to the scripts dir.