@jjagielka 2023-03-17
It makes complete sense. The package has grew really big and this PR would allow build in parts.
I've played with it a bit, and seems like a really good move.

I think we can proceed with it to be official.